# backend/src/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Library Management System API",
    "version": "1.0.0",
    "description": "Tai lieu API he thong Quan ly Thu vien (Library Management System). Ho tro xac thuc JWT, quan ly sach, muon tra, the doc gia, dat truoc, thong bao, chat truc tuyen va thong ke."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhap JWT token vao o value. Giao dien Swagger se tu dong them tien to Bearer."
      }
    }
  },
...
  "paths": {
...
    "/api/books": {
      "get": {
        "tags": ["Books"],
        "summary": "Danh sach va tim kiem sach (phan trang, loc the loai, nha xuat ban)",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Tu khoa tim kiem ten sach, tac gia, isbn" },
          { "name": "category_id", "in": "query", "schema": { "type": "string" }, "description": "Ma the loai can loc" },
          { "name": "publisher_id", "in": "query", "schema": { "type": "string" }, "description": "Ma nha xuat ban can loc" },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": { "description": "Danh sach sach va metadata phan trang" }
        }
      },
      "post": {
        "tags": ["Books"],
        "summary": "Them dau sach moi (Ho tro tu dong tao the loai & NXB neu chua co)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "isbn"],
                "properties": {
                  "title": { "type": "string", "example": "Lap trinh Flutter toan tap" },
                  "isbn": { "type": "string", "example": "9786045678901" },
                  "author": { "type": "string", "example": "Tran Van B" },
                  "category_id": { "type": "string", "example": "Cong nghe thong tin" },
                  "publisher_id": { "type": "string", "example": "NXB Thong Tin va Truyen Thong" },
                  "publication_year": { "type": "integer", "example": 2024 },
                  "description": { "type": "string", "example": "Giao trinh Flutter tu co ban den nang cao" },
                  "cover_image": { "type": "string", "example": "/uploads/covers/flutter_book.jpg" },
                  "total_copies": { "type": "integer", "example": 5 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Tao dau sach moi thanh cong" }
        }
      }
    },
...
    "/api/readers": {
      "get": {
        "tags": ["Readers"],
        "summary": "Danh sach doc gia (tim kiem, loc khoa phong, trang thai the - Thu thu/Admin)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "department", "in": "query", "schema": { "type": "string" } },
          { "name": "card_status", "in": "query", "schema": { "type": "string", "enum": ["active", "suspended", "expired"] } }
        ],
        "responses": {
          "200": { "description": "Danh sach ho so doc gia" }
        }
      },
      "post": {
        "tags": ["Readers"],
        "summary": "Tao ho so doc gia va cap the thu vien moi (Thu thu/Admin)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["full_name", "email"],
                "properties": {
                  "full_name": { "type": "string", "example": "Tran Thi Mai" },
                  "email": { "type": "string", "example": "mai.tran@example.com" },
                  "phone": { "type": "string", "example": "0901234567" },
                  "address": { "type": "string", "example": "Quan Cau Giay, Ha Noi" },
                  "student_id": { "type": "string", "example": "SV20260199" },
                  "department": { "type": "string", "example": "Cong nghe thong tin" },
                  "card_expiry_date": { "type": "string", "format": "date", "example": "2027-12-31" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Tao ho so doc gia thanh cong" }
        }
      }
    },
...
    "/api/holds": {
      "get": {
        "tags": ["Holds"],
        "summary": "Xem toan bo danh sach dat truoc sach (Thu thu/Admin)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Danh sach dat truoc" }
        }
      },
      "post": {
        "tags": ["Holds"],
        "summary": "Dat truoc mot cuon sach (khi khong con ban sao san sang)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["book_id"],
                "properties": {
                  "book_id": { "type": "string", "example": "book-uuid-1" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Dat truoc sach thanh cong" },
          "400": { "description": "Sach van con ban sao san sang hoac da dat truoc roi" }
        }
      }
    },
...
    "/api/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Lay danh sach the loai sach",
        "responses": {
          "200": { "description": "Danh sach the loai" }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Them the loai sach moi (Thu thu/Admin)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["category_name"],
                "properties": {
                  "category_name": { "type": "string", "example": "Khoa hoc may tinh" },
                  "description": { "type": "string", "example": "Sach ve tri tue nhan tao, du lieu lon va lap trinh" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Them the loai thanh cong" }
        }
      }
    },
...
    "/api/publishers": {
      "get": {
        "tags": ["Publishers"],
        "summary": "Lay danh sach nha xuat ban",
        "responses": {
          "200": { "description": "Danh sach nha xuat ban" }
        }
      },
      "post": {
        "tags": ["Publishers"],
        "summary": "Them nha xuat ban moi (Thu thu/Admin)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["publisher_name"],
                "properties": {
                  "publisher_name": { "type": "string", "example": "Nha xuat ban Tre" },
                  "address": { "type": "string", "example": "161B Ly Chinh Thang, Quan 3, TP.HCM" },
                  "phone": { "type": "string", "example": "02839316289" },
                  "email": { "type": "string", "example": "hopthu@nxbtre.com.vn" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Them nha xuat ban thanh cong" }
        }
      }
    },
...
    "/api/shelves": {
      "get": {
        "tags": ["Shelves"],
        "summary": "Lay danh sach toan bo ke sach",
        "responses": {
          "200": { "description": "Danh sach ke sach" }
        }
      },
      "post": {
        "tags": ["Shelves"],
        "summary": "Them ke sach moi (Thu thu/Admin)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["shelf_name", "location"],
                "properties": {
                  "shelf_name": { "type": "string", "example": "Ke A1 - Tang 2" },
                  "location": { "type": "string", "example": "Khu vuc Sach Khoa hoc Tu nhien" },
                  "capacity": { "type": "integer", "example": 100 },
                  "description": { "type": "string", "example": "Ke kim loai 4 tang" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Them ke sach thanh cong" }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Lay danh sach thong bao cua doc gia",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Danh sach thong bao" }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Xem danh sach tai khoan he thong",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "role", "in": "query", "schema": { "type": "string", "enum": ["admin", "librarian", "reader"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Danh sach tai khoan" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Tao tai khoan moi boi Admin (vi du: Tao thu thu moi)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password", "email", "role"],
                "properties": {
                  "username": { "type": "string", "example": "librarian2" },
                  "password": { "type": "string", "example": "MatKhau123@" },
                  "email": { "type": "string", "example": "librarian2@library.com" },
                  "role": { "type": "string", "enum": ["librarian", "admin", "reader"], "default": "librarian" },
                  "full_name": { "type": "string", "example": "Le Van Cuong" },
                  "phone": { "type": "string", "example": "0912345678" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Tao tai khoan thanh cong" }
        }
      }
    },
...
  }
}